+++ include/peer_manager.hpp
@@ -35,6 +35,9 @@ public:
     /** @copydoc IPeerManager::get_offline */
     etl::vector<NodeId, MAX_PEERS> get_offline(int64_t now_ms) override;
 
+    /** @copydoc IPeerManager::is_online */
+    bool is_online(NodeId id, int64_t now_ms) const override;
+
     /** @copydoc IPeerManager::update_last_seen */
     void update_last_seen(NodeId id, int64_t now_ms) override;